+++ shaders/main_files/gbuffers_hand.glsl
@@ -96,6 +96,7 @@ void main() {
 void main() {
 	texcoord = (gl_TextureMatrix[0] * gl_MultiTexCoord0).xy;
 	lmcoord  = (gl_TextureMatrix[1] * gl_MultiTexCoord1).xy;
+	adjustLmcoord(lmcoord);
 	
 	
 	#ifdef ISOMETRIC_RENDERING_ENABLED